move MenuContext to components
