+++ src/app/layout.tsx
@@ -6,7 +6,7 @@ import { ApolloWrapper } from "common/apollo/apollo-wrapper";
 import Footer from "common/components/Footer";
 import Header from "common/components/Header/Header";
 import { OpenEntitiesContextProvider } from "common/components/EntityDetails/OpenEntitiesTabs/OpenEntitiesContext";
-import { MenuControlProvider } from "common/MenuContext";
+import { MenuControlProvider } from "common/components/MenuContext";
 import MuiXLicense from "common/components/MuiXLicense";
 import { Suspense } from "react";
 import { theme } from "./theme";
